De-parameterize `PacketError` type.
This is the first step in simplifying the error types. The payload encode methods that returned a
generic `PayloadError` now return a concrete type, that can easily be converted into a
`PacketError`. The `PacketError` enum gained variants and `impl From<>` for all those.

I couldn't get the generic `impl<T: Packet + fmt::Debug + 'static> Encodable for T` to compile
(something about Struct->Trait->Trait->Method being harder to resolve when said Struct is in an enum
variant), but using the macro to generate the equivalent works nicely.

This commit breaks the API for users that looked into the error type instead of simply bailing
out. Due to the difficulty of looking into this type, I doubt that anybody was doing this correctly
anyway, but this is nevertheless an API change to ducument.