Remove `VariablePacketError` type.
Use `PacketError` instead. Yay to code deduplication :)

To keep the functionality as-is, I added a `Vec<u8>` to `FixedHeaderError::Unrecognized/Reserved`,
that holds the remaining bytes. Consuming those bytes doesn't seem to be done consistently (for
example it's only done if you decode as a `VariablePacket`), and no unittest take notice of the
issue. Will need to take a second look at those later.