Refactor the error types.
The general philosophy was to consistently bubble common errors like `io::Error` and
`StringEncodeError` up to the parent error, eventually reaching `PacketError`. In doing so, various
error types and variants naturally disappeared, the code got simpler, and some unused enum variants
were deleted.

A few error variants/enum were documented, and a `errotypes.dot` file was added to document the
current error graph.

This scratches my original itch of easily recognizing "more bytes needed to finish parsing" errors,
but it does lose some information like "what packet type was being treated when we reached an
utf8-encoding error ?". I've tried various approaches, but it's tricky to get working without
relying on some error-chaining crate like `snafu` or `failure`, which might not be acceptable for
this lib (and I don't think we want to go back to a `PacketError<Packet>` type). Losing that context
is IMHO ok (there's not a lot that can be done with the context anyway), but a survey of usecases
would be nice to confirm that opinion.