Simplify `FixedHeaderError`.
* Brought `PacketTypeError::InvalidFlag` directly into `FixedHeaderError`.
* No longer attempting to read the rest of the packet when a `FixedHeaderError` happens.
* Expanded `FixedHeaderError` documentation.

The second bullet point might be controversial as it removes a feature that might be in use. But
reading the rest of the packet is a spec violation (user should close the connection instead), it
degrades performance, complicates code, and seems like a risky bet whichever way you look at it. So
I left the parsed type/flag/lenght in place in case some adventurous user wants to read/parse the
rest of the stream, but made it clear that the user is on his own if he does that.